requirements.txt
# Core AI and Analysis
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-groq>=0.0.1
openai>=1.6.0
groq>=0.4.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
pylint>=3.0.0

# CLI and Utilities
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.25.0

# Vector Database for RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1


# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.6.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0


